+++ rpc/client/client.go
@@ -10,6 +10,8 @@ import (
 	"encoding/hex"
 	"time"
 
+	"github.com/33cn/chain33/common/crypto"
+
 	"github.com/33cn/chain33/system/address/btc"
 
 	"github.com/33cn/chain33/system/crypto/btcscript"
@@ -513,7 +515,24 @@ func (c *ChannelClient) SignWalletRecoverTx(req *types.ReqSignWalletRecoverTx) (
 	return reply, nil
 }
 
-//func (c *ChannelClient)GetHighestBlockNum(param *types.ReqNil)(*types.ReplyBlockHeight,error){
-//
-//	return c.GetHighestBlockNum(param)
-//}
+
+// GetCryptoList 获取加密算法列表
+func (c *ChannelClient) GetCryptoList() *types.CryptoList {
+	names, ids := crypto.GetCryptoList()
+	list := &types.CryptoList{Cryptos: make([]*types.Crypto, len(names))}
+	for i, name := range names {
+		list.Cryptos[i] = &types.Crypto{Name: name, TypeID: ids[i]}
+	}
+	return list
+}
+
+// GetAddressDrivers 获取已注册地址插件
+func (c *ChannelClient) GetAddressDrivers() *types.AddressDrivers {
+	drivers := address.GetDriverList()
+	list := &types.AddressDrivers{Drivers: make([]*types.AddressDriver, 0, len(drivers))}
+	for id, driver := range drivers {
+		list.Drivers = append(list.Drivers, &types.AddressDriver{Name: driver.GetName(), TypeID: id})
+	}
+	return list
+}
+