+++ CHANGELOG.md
@@ -1,5 +1,22 @@
 changelog
 
+<a name="1.67.1"></a>
+## [1.67.1](https://github.com/33cn/chain33/compare/v1.67.0...v1.67.1) (2022-03-28)
+
+
+### Bug Fixes
+
+* fix 32 bit machine build ([a494d18](https://github.com/33cn/chain33/commit/a494d18))
+* update patch version ([c0c2d0b](https://github.com/33cn/chain33/commit/c0c2d0b))
+
+<a name="1.67.0"></a>
+# [1.67.0](https://github.com/33cn/chain33/compare/v1.66.5...v1.67.0) (2022-03-21)
+
+
+### Features
+
+* add multiple address format support(#1181) ([659f342](https://github.com/33cn/chain33/commit/659f342)), closes [#1181](https://github.com/33cn/chain33/issues/1181)
+
 <a name="1.66.5"></a>
 ## [1.66.5](https://github.com/33cn/chain33/compare/v1.66.4...v1.66.5) (2022-02-18)